management_app/requirement.txt
"celery[redis]"
django-debug-toolbar==3.6.0
redis==4.3.4


